drm/docs: Include hdmi infoframe helper reference
Thierry created such nice kerneldocs, it's a shame we've left them
lingering!

For the fun of it also add a bit of kerneldoc to the header so that we
can also include that. Just in case someone adds kerneldoc in there.

Cc: Thierry Reding <thierry.reding@gmail.com>
Reviewed-by: Alex Deucher <alexander.deucher@amd.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit 2d123f463669cb7b84b56aa00e073ce07fe7aff2)

Signed-off-by: Dheeraj Jamwal <dheerajx.s.jamwal@intel.com>