drm/i915: add gen-specific runtime suspend/resume functions
We're adding runtime suspend support to more platforms, so organize
the code in a way that all a new platform needs to do is to add its
own gen-specific functions. Also rename the i915_ functions to intel_
to make it clear that it's the top level one, not something that just
runs on i915 platforms.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>
Reviewed-by: Imre Deak <imre.deak@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit 97bea20794f600ae9248bef296e3b62d12087a1e)

Signed-off-by: Dheeraj Jamwal <dheerajx.s.jamwal@intel.com>